+++ src/main/kotlin/be/mgx/networks/BasicAndPerceptron.kt
@@ -5,16 +5,11 @@ import be.mgx.core.math.Matrix
 import be.mgx.functions.ActivationFunction
 import be.mgx.util.*
 import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.csv.Csv
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.decodeFromStream
 import kotlinx.serialization.json.encodeToStream
 import picocli.CommandLine.*
 import java.io.File
-import java.io.FileInputStream
 import java.io.FileOutputStream
-import java.nio.charset.StandardCharsets
 
 @OptIn(ExperimentalSerializationApi::class)
 @Command(name = "logical-and", abbreviateSynopsis = true)
@@ -47,12 +42,12 @@ class BasicAndPerceptron : IAbstractNetwork {
         var (X, Y) = model.map { it.chunked(2) }
             .transpose()
 
-        X = X.map { x -> listOf(1f) + x } // Prepend 1f to make the matrix a 1x3 input matrix
+        X = X.map { x -> listOf(1.0) + x } // Prepend 1.0 to make the matrix a 1x3 input matrix
 
         network.train(
             X.map { x -> Matrix.createMatrix(1, 3) { x } },
             Y.map { y -> Matrix.createMatrix(1, 1) { y } },
-            1f,
+            1.0,
             ErrorFunctions.basicErrorFunction(),
             StopFunctions.iterationStopFunction(6),
             listOf(saveLayerWeights)
@@ -60,7 +55,7 @@ class BasicAndPerceptron : IAbstractNetwork {
 
         saveNetworkModelToFile(network, modelFile)
 
-        val inputs = mutableListOf<ArrayList<Float>>()
+        val inputs = mutableListOf<ArrayList<Double>>()
 
         for (arrayList in model) {
             val newArrayList = arrayListOf(arrayList[0], arrayList[1])
@@ -82,7 +77,7 @@ class BasicAndPerceptron : IAbstractNetwork {
         @Option(names = ["-m", "--model"])
         model: File,
         @Parameters(description = ["Inputs (separated by commas) for the neural network"], split = ",")
-        inputs: List<Float>
+        inputs: List<Double>
     ): Int {
         val network = retrieveNetworkModelFromFile(model)
         val input = Matrix.createMatrix(1, inputs.size) { inputs }
